// temper/game-combat-addon/src/core/types.ts
export type CoreLogLine = (number | string | undefined)[]
// ...
export interface CurrentData extends Partial<FightRecapData>, Partial<GroupRecapData> {
  log: CoreLogLine[]
  DPSOut: number
  DPSIn: number
  HPSOut: number
  HPSAOut: number
  HPSIn: number
  dpstime: number
  hpstime: number
  groupDPSOut: number
  groupDPSIn: number
  groupHPSOut: number
  groupHPS: number
  units?: Record<number, UnitEntry>
}